asm: fix SSE2 feature detection for existing operations (#10711)
Previously, these `*pd` vector operations had been classified as SSE
when in fact they are SSE2. This is not the most critical bug fix, since
SSE2 was released in 2000 and all x86 CPUs past that date should have
both SSE and SSE2 present. But this corrects the record and illustrates
how to add new features.