pulley: Allow disabling SIMD in the interpreter at compile-time
This commit adds a new feature to Pulley which is used to reduce the
compiled code size of the interpreter itself by disabling SIMD opcode
interpretation at compile-time. The goal here is to be low-impact on
Pulley itself to avoid needing `#[cfg]` all over the place and to
additionally avoid the need to redesign Pulley's opcode macro for use in
various parts of Wasmtime.

Methods are annotated with a custom macro in the interpreter which
registers a `#[cfg]` that either does the listed implementation or
switches to an implementation that emits a trap if executed. This means
that it's safe to execute mismatched code where SIMD was enabled at
compile time but disabled at runtime, it just means the semantics may be
a bit surprising to debug.

Note that this SIMD is still enabled by default, and an explicit `--cfg`
via `RUSTFLAGS` is required to compile-out the SIMD support. Cargo
features aren't a great fit for this sort of feature so an explicit flag
is used.