Inline assembler-x64 `generated_files` in `main.rs`
The public function `generated_files` in `cranelift-assembler-x64` makes
the generated `rlib` non-deterministic because it contains the full
paths of generated files. But this function is only used in `main.rs` of
the same crate, so this change inlines it there to keep the library artifact
deterministic while maintaining the same behavior.