Fix another libcall popping up with simd (#10735)
This commit is similar to #10699, another instance of a libcall popping
up late in the x64 backend. Fuzzing found this issue and to help verify
this is the last one I've run the whole `*.wast` test suite with the
x86_64 baseline (no target features) and saw the panic before this PR
and no more panics after.