[33.0.0] Backport some fixes from main (#10743)
* Skip a test with threads on ASAN (#10728)

It's expected that this has memory leaks, there's no clean shutdown in
the CLI right now.

* Fix missing libcalls with simd float rounding (#10699)

This commit fixes some more fallout found on oss-fuzz about the x64
generating rounding builtins when it shouldn't be. This situation is
caused by simd float rounding instructions which the x64 backend lowers
to libcall-per-element and now needs to move that logic to the frontend
instead.

* Fix another libcall popping up with simd (#10735)

This commit is similar to #10699, another instance of a libcall popping
up late in the x64 backend. Fuzzing found this issue and to help verify
this is the last one I've run the whole `*.wast` test suite with the
x86_64 baseline (no target features) and saw the panic before this PR
and no more panics after.

* Inline assembler-x64 `generated_files` in `main.rs` (#10739)

The public function `generated_files` in `cranelift-assembler-x64` makes
the generated `rlib` non-deterministic because it contains the full
paths of generated files. But this function is only used in `main.rs` of
the same crate, so this change inlines it there to keep the library artifact
deterministic while maintaining the same behavior.

---------

Co-authored-by: Adam Bratschi-Kaye <adam.bratschikaye@dfinity.org>