Cranelift: update to regalloc2 0.12.2; support many return values. (#10747) (#10748)
* Cranelift: update to regalloc2 0.12.2; support many return values.

Prior versions of regalloc2 could not support more than 255 operands on
an instruction, and together with the integrated return-value loads on
call instructions introduced in #10502, this caused issues with calls
with many returns. This PR upgrades to a version of RA2 that supports up
to `2^16 - 1` operands per instruction (well in excess of the maximum of
1000 return/result values per Wasm's implementation limits, for
example).

Fixes #10741.

* Update vets

---------

Co-authored-by: Alex Crichton <alex@alexcrichton.com>