winch(aarch64): Run wast test in CI
Closes https://github.com/bytecodealliance/wasmtime/issues/9566

This commit takes inspiration from
https://github.com/bytecodealliance/wasmtime/pull/10738/files

During the 2025-05-07 Cranelift meeting we discussed potential avenues
to start running wast tests for winch-aarch64. One potential idea was
to declare an allowlist of tests that should be exectued and
potentially ignore everyhing else. Although that idea works, I decided
to diverge a bit from it, in favor of introducing a very strict
classification of the state of tests for aarch64, namely, in this
commit tests are classified as:

* Run and expected to pass
* Run and exepected to fail in a recoverable way
* Don't run since it's known that they produce an unrecoverable error
e.g., segafault.

This approach is probably more verbose than the one discussed in the
meeting, however, I think it's easier to have a global view of that
status for aarch64 and potentially other backends in the future.