winch(aarch64): Run wast test in CI (#10750)
* winch(aarch64): Run wast test in CI

Closes https://github.com/bytecodealliance/wasmtime/issues/9566

This commit takes inspiration from
https://github.com/bytecodealliance/wasmtime/pull/10738/files

During the 2025-05-07 Cranelift meeting we discussed potential avenues
to start running wast tests for winch-aarch64. One potential idea was
to declare an allowlist of tests that should be exectued and
potentially ignore everyhing else. Although that idea works, I decided
to diverge a bit from it, in favor of introducing a very strict
classification of the state of tests for aarch64, namely, in this
commit tests are classified as:

* Run and expected to pass
* Run and exepected to fail in a recoverable way
* Don't run since it's known that they produce an unrecoverable error
e.g., segafault.

This approach is probably more verbose than the one discussed in the
meeting, however, I think it's easier to have a global view of that
status for aarch64 and potentially other backends in the future.

* Formatting

* Extract common tests that fail in all targets

* Move the `Compiler::should_fail` call at runtime

`Compiler::should_fail` makes use of `#[cfg(target_arg="...")]` directives to
derive the compiler-arch-Wasm proposal support matrix, however,
`Compiler::should_fail` is also called from the `wasmtime_test` macro
which introduces conflicts with the `target_arch` resolution (within macros,
these directives resolve to the host compiler, instead of the target).

This commit emits `Compiler::should_fail` at runtime instead of at
compile time, which fixes the issue mentioned above.

* Use `.unwrap` when test is expected to pass