x64: refactor the representation of `*Mem` and `*MemImm`
While reflecting on why #10762 is hard, it seemed that the underlying
problem to many ISLE matching issues is our inability to actually
inspect the variants of the register specific versions of `RegMem` and
`RegMemImm`. We had been wrapping instances of these `Reg*` types, but
this change pushes the variants down to: `GprMem`, `GprMemImm`,
`XmmMem`, `XmmMemAligned`, `XmmMemImm`, and `XmmMemAlignedImm`. With
this, it should be possible to match directly on the variants in ISLE.
This does not change any functionality.