mpk: fix example CLI argument (#10777)
The `--memory-size` argument caused failures when running the `mpk.rs`
example. This was due to some `clap`-related conversions that no longer
worked as when this was written:

```
thread 'main' panicked at examples/mpk.rs:69:18:
Mismatch between definition and access of `memory_size`. Could not
downcast to usize, need to downcast to u64
```

This change explicitly uses `u64` for the CLI argument, fixing the
failure.