Add a `--generate-dwarf[=true|false]` flag to the `wast` subcommand
This commit adds support for the `wast` subcommand to generate DWARF
debugging information for the text-to-binary translation that happens.
This leans on the support within the `wast` crate already and enables
more easily seeing where a trap is happening during development of
tests due to the fact that line numbers are available as opposed to just
having binary offsets with a stack trace.

Currently this is enabled-by-default with an option to disable it, but
I'm not actually sure if we have a use case for disabling it...