Classify Winch with Tier 1 support (#10755)
* Classify Winch with Tier 1 support

This commit classifies the Winch compiler in our documentation as a Tier
1 feature. This reflects the degree of work and investment put into it
as well as the thorough quality vetting that has already happened in
production. For Wasmtime this means that bugs in Winch are candidates
for security releases as well.

This additionally refines our tiers-of-support documentation to reflect
how not all compilers are required to support all features. Tables are
added of per-architecture support for each compiler as well as refined
wording of the various requirements of what is required for each tier as
well.

* Re-organize some docs