x64: Stringify fixed registers with their runtime value (#10815)
This commit updates the display implementation of instructions using
fixed registers to use the current value of the register rather than the
end-result that'll happen after register allocation. That should help
ensure that if a virtual register is being used in an instruction
that'll get printed instead of the fixed value that should pop out
after regalloc. Disassembly shouldn't change though after register
allocation because the value of the register should match the fixed
encoding.