x64: Migrate `SignExtendData` to new assembler (#10814)
* x64: Migrate `SignExtendData` to new assembler

I woke up early today and figure how better to start the day than to
frob some assembler instructions. I'd never seen the `SignExtendData`
before and it was trickier and more subtle than I thought. The end
result of this PR is to remove this pseudo-instruction and replace it
with the per-instruction equivalents that it can generate.

The first thing surprising about this instruction is that the Capstone
and Intel names disagree. For example what Intel calls `CBW` Capstone
calls `cbtw`. This is true for all the various instructions here so the
names are all changing subtly.

The second surprising thing was that the behavior of the instruction
depended on the operand size. This instruction was tasked with
sign-extending data just before a `div` instruction which requires the
operand to be in `AX` (sign-extended) for 8-bit division, but otherwise
requires `DX:AX` or wider for wider division. That meant that the 8-bit
version of `SignExtendData` would read `AL` and write `AX`, but the
wider variants would read `AX` and write `DX` for example.

This involved writing a specialized helper in ISLE for handling the
16-bits-or-higher width and otherwise customizing each of the various
cases for 8-bit or wider already in ISLE. Winch received appropriate
minor updates as well, and in the end we've now got a few more
instructions bound which Cranelift doesn't currently used but should be
easier to do so in the future.

* Update cranelift/assembler-x64/meta/src/instructions/bitmanip.rs

Co-authored-by: Andrew Brown <andrew.brown@intel.com>

---------

Co-authored-by: Andrew Brown <andrew.brown@intel.com>