x64: Migrate `ShiftR` pseudo-instruction to the new assembler
Shifts and rotates were all represented with `ShiftR` and they're now
all represented with individual instructions. Additionally the old
`Imm8{Gpr,Reg}` abstractions were largely removed and/or replace to only
be helpers in ISLE.

This does regress pcc a bit more than it already is at due to the shift
instructions not being special-cased for pcc processing.