winch: Use innermost frame to emit `br_table` (#10821)
* winch: Use innermost frame to emit `br_table`

Prior to this commit, the lowering of br_table used the default target
as the relative reference to pop ABI results. This approach is not
ideal, as it leads to a loss of precision regarding the stack pointer
position, causing unmet invariants at jump sites for all other
branches.

This commit, instead, takes advantage of the fact that all jumps are
"outward" and uses the innermost frame to pop ABI results when
lowering br_table. This ensures two main things:

* The stack pointer offset is correctly positioned according to the
  expectations of the innermost block's end sequence, which will be
  handled in the next eventual `end` sequence.
* We meet the jump site invariants introduced by CodegenContext::br,
  which take advantage of Wasm semantics, given that all jumps are
  "outward".

* Format