winch: Crank the ratchet on tests that can fail
This commit moves all "ignore these outright" tests for Winch and
AArch64 to "run the test and expect failure" after the fix in #10826
seems to have fixed many of them. Two failures remained which were easy
enough to fix here:

* The `cmov` implementation only worked on integer registers, not
  floating-point registers, leading to a panic. Some small instruction
  helpers were added for floating-point selects to resolve this panic
  and get some tests passing.

* More subtly the calculation of a table's function address was
  calculating the wrong value on AArch64. This happened because the
  `mul` low-level helper on AArch64 required use of the `scratch`
  register when an immediate was an operand, unlike on x86_64. The table
  address helper was also using the `scratch` register for other data,
  however, which meant that loading an immediate into `scratch`
  clobbered the previous data. The fix here was to allocate a temporary
  register in the function address routine.

After these fixes the `ignore` method is no longer necessary as all
tests can be run and be flagged as either pass or fail. This
additionally adds assertions in the `asm.rs` layer that any usage of
`regs::scratch()` has an assertion that the other operand isn't also the
scratch register to help prevent this issue from cropping back up. This
identified one other location in fuel handling where the scratch
register collided.