x64: Convert some lane-manipulation instructions to the new assembler (#10825)
* x64: Convert some lane-manipulation instructions to the new assembler

This was mostly straightforward but ended up requiring handling for a
new style of instruction not previously before-seen: a write-only
`GprMem` operand. This needed some extra treatment/helpers in the ISLE
integration to ensure that this was all hooked up correctly from ISLE
constructors down to the assembler instructions.

* Remove `GprToXmm` as well

Already have mappings for `movd` and `movq`, so let's migrate this as well.

* Remove some old emit tests