asm: model a way to modify the lower three opcode bits (#10832)
* asm: model a way to modify the lower three opcode bits

The x86 ISA includes a special format that, for certain instructions,
writes the low three bits of the destination register into the low three
bits of the opcode. This is specified with `+rb`, `+rw`, `+rd`, and
`+ro` in the reference manual. This change adds a way to model this when
defining instructions, e.g., `rex(...).rb()`.

* review: refactor using `OpcodeMod`, add test

* fix: convert `bits` into its new `u16` type