x64: Update codegen of `XmmCmove` pseudo-inst (#10839)
In #4317 this instruction was updated to handle 128-bit vectors in
addition to the previous handling of 32/64-bit floats. Originally the
pseudo-instruction used `movs{s,d}` to achieve its task and when adding
128-bit support I mistakenly switched both f32/f64 paths to using
`movsd` instead of conditionally using `movss` for `f32`. In retrospect
though it's probably best to use a full register move here instead of
just a singular mov because `movss` and `movsd` preserve the upper bits
of the register, needlessly creating a data dependency with the previous
value in the register.

This commit updates this helper to using `Inst::gen_move` which already
internally does this optimization of using `movaps`, a documented
zero-latency instruction, for all xmm-style register movements.