Merge `vm::Instance::vmctx_plus_offset{,_mut}`
This commit merges two helper methods to calculate an internal pointer
with a `VMContext`. Historically these two methods mapped different
levels of mutability but nowadays there's no need for this. The
provenance/mutability of `VMContext` memory is independent of `&self`
meaning that it's safe to create/mutate through a `NonNull` in the vmctx
area through a `&self`-derived pointer.

The purpose of this commit is to avoid needlessly requiring `&mut self`
throughout methods on `Instance`. Currently many of them require
mutability to only create a pointer, and this mutability is a burden
that need not be required.